Restructure Test Client, Testing Ctd
Restructured the test client to be a express server that serves the
public files instead of using http-server client due to caching issues
Room Socket Events mostly work, though creating doesnt work fully as
we get back a null value instead of the listing. So will have to work
on the timing of the functions called.

Thats the commit for today =)